+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog
 
 - Drop json support, adopt yaml only format for scenarios (#5)
 
+- Drop ini format for metadata files, use YAML (#5)
+
 - `play.execute` no more accepts raw data string), consumes a list of commands.
   Introduced `play.execute_raw` accepting raw data string.